pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
flask-restful = "^0.3.9"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
mysqlclient = "^2.1.1"
flask-cors = "^3.0.10"
bcrypt = "^4.0.1"
flask-jwt-extended = "^4.4.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
